requirements.txt
requests==2.22
urllib3==1.25.6
PIL==1.1.6
python-dotenv==0.10.3
instabot==0.4.5